feat: 迁移部分float类型数据到double, 修改注入器使其与实现解耦 fix: 修复一堆测试

+++ Allay-API/src/main/java/cn/allay/api/block/component/impl/attribute/BlockAttributes.java
@@ -36,7 +36,7 @@ public class BlockAttributes {
             .create();
 
     protected static AABB parseAABBStr(String str) {
-        var numbers = StringUtils.fastSplit(str, ",").stream().map(Float::valueOf).toList();
+        var numbers = StringUtils.fastSplit(str, ",").stream().map(Double::valueOf).toList();
         return AABB.of(numbers.get(0), numbers.get(1), numbers.get(2), numbers.get(3), numbers.get(4), numbers.get(5));
     }
     @Builder.Default
